+++ test/integration/replicaset/test_replicaset_promote.py
@@ -150,7 +150,7 @@ def test_promote_cconfig_failovers(
             box_ctl_promote(tt_cmd, app_name, "election-failover-1", tmpdir)
 
         if stop_inst:
-            stop_cmd = [tt_cmd, "stop", f"{app_name}:{stop_inst}"]
+            stop_cmd = [tt_cmd, "stop", "-y", f"{app_name}:{stop_inst}"]
             rc, _ = run_command_and_get_output(stop_cmd, cwd=tmpdir)
             assert rc == 0