+++ test/integration/replicaset/test_replicaset_roles_add.py
@@ -162,7 +162,7 @@ def test_replicaset_cconfig_roles_add(
         start_application(tt_cmd, tmpdir_with_cfg, app_name, instances)
 
         if stop_instance:
-            stop_cmd = [tt_cmd, "stop", f"{app_name}:{stop_instance}"]
+            stop_cmd = [tt_cmd, "stop", "-y", f"{app_name}:{stop_instance}"]
             rc, _ = run_command_and_get_output(stop_cmd, cwd=tmpdir_with_cfg)
             assert rc == 0
         if is_add_role: