+++ test/integration/running/test_running_cluster.py
@@ -26,7 +26,7 @@ def start_application(cmd, workdir, app_name, instances):
 
 
 def stop_application(tt_cmd, app_name, workdir, instances):
-    stop_cmd = [tt_cmd, "stop", app_name]
+    stop_cmd = [tt_cmd, "stop", "-y", app_name]
     stop_rc, stop_out = run_command_and_get_output(stop_cmd, cwd=workdir)
     assert stop_rc == 0